+++ example/lib/presentation/home_page.dart
@@ -1,5 +1,3 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:helium_flutter/helium_flutter.dart';
 import 'package:helium_flutter_example/presentation/revenue_cat_page.dart';
@@ -86,12 +84,10 @@ class _HomePageState extends State<HomePage> {
             ElevatedButton(
               key: ValueKey('present_upsell'),
               onPressed: () async {
-                final presentResult = await _heliumFlutterPlugin.presentUpsell(
-                      trigger: 'onboarding',
-                      context: context,
-                    ) ??
-                    '';
-                log("presentUpsell result - $presentResult");
+                await _heliumFlutterPlugin.presentUpsell(
+                  trigger: 'onboarding',
+                  context: context,
+                );
               },
               child: Text('Present upsell'),
             ),