+++ lib/core/helium_callbacks.dart
@@ -1,9 +1,19 @@
 import 'package:helium_flutter/types/helium_transaction_status.dart';
 
 abstract class HeliumCallbacks {
-  Future<HeliumTransactionStatus> makePurchase(String productId);
+  Future<HeliumPurchaseResult> makePurchase(String productId);
 
-  Future<bool> restorePurchases(bool status);
+  Future<bool> restorePurchases();
 
   Future<void> onPaywallEvent(Map<String, dynamic> heliumPaywallEvent);
+}
+
+class HeliumPurchaseResult {
+  final HeliumTransactionStatus status;
+  final String? error;
+
+  HeliumPurchaseResult({
+    required this.status,
+    this.error,
+  });
 }